+++ src/TheNerdCollective.Blazor.ThemeKit/ServiceCollectionExtensions.cs
@@ -1,4 +1,5 @@
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection.Extensions;
 
 namespace TheNerdCollective.Blazor.ThemeKit;
 
@@ -14,7 +15,7 @@ public static class ServiceCollectionExtensions
         }
 
         services.AddSingleton<MudThemePreferencesService>();
-        services.AddSingleton<IThemeJsonFilePersistence, NullThemeJsonFilePersistence>();
+        services.TryAddSingleton<IThemeJsonFilePersistence, NullThemeJsonFilePersistence>();
         services.AddScoped<IMudThemeSessionStore, MudThemeSessionStore>();
         services.AddScoped<IMudThemeStateService, MudThemeStateService>();
         services.AddSingleton<IThemeEditorGate, ThemeEditorGate>();